// src/app/api/model/alternativeNameList.ts
/**
 * TMDB API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { AlternativeName } from './alternativeName';


export interface AlternativeNameList { 
    id?: number;
    results?: Array<AlternativeName>;
}