// src/app/api/model/networkDetails.ts
/**
 * TMDB API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface NetworkDetails { 
    headquarters?: string;
    homepage?: string;
    id?: number;
    logo_path?: string | null;
    name?: string;
    origin_country?: string;
}